Separate datatypes out into their own file.
This simplifies the first setup of Galaxy-M but more importantly every subsequent update.  Deployers will not have to merge these changes with each update to Galaxy's copy of binary.py.
